Moved anonymous class in NNRankTrainer to org.ltr4l.nn.NNMLP
