+++ src/main/java/org/ltr4l/trainers/NNRankTrainer.java
@@ -17,12 +17,8 @@
 package org.ltr4l.trainers;
 
 import java.io.Reader;
-import java.util.List;
 
-import org.ltr4l.nn.Activation;
-import org.ltr4l.nn.MLP;
-import org.ltr4l.nn.NetworkShape;
-import org.ltr4l.nn.Optimizer;
+import org.ltr4l.nn.*;
 import org.ltr4l.query.Document;
 import org.ltr4l.query.Query;
 import org.ltr4l.query.QuerySet;
@@ -45,7 +41,7 @@ public class NNRankTrainer extends MLPTrainer<MLP> {
   }
 
   @Override
-  protected MLP constructRanker(){
+  protected NNMLP constructRanker(){
     int featureLength = trainingSet.get(0).getFeatureLength();
     //Add an output layer with number of nodes equal to number of classes/relevance categories.
     NetworkShape networkShape = config.getNetworkShape();
@@ -54,19 +50,7 @@ public class NNRankTrainer extends MLPTrainer<MLP> {
     Optimizer.OptimizerFactory optFact = config.getOptFact();
     Regularization regularization = config.getReguFunction();
     String weightModel = config.getWeightInit();
-    return new MLP(featureLength, networkShape, optFact, regularization, weightModel) {
-      @Override
-      public double predict(List<Double> features) {
-        double threshold = 0.5;
-        forwardProp(features);
-        for (int nodeId = 0; nodeId < network.get(network.size() - 1).size(); nodeId++) {
-          MNode node = network.get(network.size() - 1).get(nodeId);
-          if (node.getOutput() < threshold)
-            return nodeId - 1;
-        }
-        return network.get(network.size() - 1).size() - 1;
-      }
-    };
+    return new NNMLP(featureLength, networkShape, optFact, regularization, weightModel);
 
   }
 
@@ -93,4 +77,5 @@ public class NNRankTrainer extends MLPTrainer<MLP> {
     }
     if (batchSize != 0) ranker.updateWeights(lrRate, rgRate);
   }
+
 }
